4.1 Opengl shaders seem to work after a long time of debugging
Need to define gl_Position still in vertex shader even if the output
is defined. Next going to test if the test shader works, which is
now commented out. Really crazy amount of debugging and I don't
even remember all the things I had to do to get this working.